[project @ 5ba2296c5933d44872cbfb95eb92dc6cba3ce6a2] Changed some of the casts to make it compile for me. --Thotter
Original author: w4dimenscor <w4dimenscor>
Date: 2007-01-26 16:49:57+00:00

git-svn-id: svn+ssh://4dimensions.org/home/game-source/repositories/gamesrc@304 254f5b4c-bdf4-4503-907e-934c7b59e499

+++ dg_variables.c
@@ -4,8 +4,8 @@
 *                                                                         *
 *                                                                         *
 *  $Author: w4dimenscor $         		                          *
-*  $Date: 2006/11/11 21:16:52 $                                           * 
-*  $Revision: 1.39 $                                                      *
+*  $Date: 2007/01/26 16:49:57 $                                           * 
+*  $Revision: 1.40 $                                                      *
 **************************************************************************/
 
 #include "conf.h"
@@ -189,7 +189,7 @@ int text_processed(char *field, char *subfield, struct trig_var_data *vd,
     //char tmpvar[MAX_STRING_LENGTH];
 
     if (!str_cmp(field, "strlen")) {                     /* strlen    */
-        snprintf(str, slen, "%d", vd->value.length());
+        snprintf(str, slen, "%d", (int)(vd->value.length()));
         return TRUE;
     } else if (!str_cmp(field, "trim")) {                /* trim      */
 #if 0
